backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain and RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
faiss-cpu==1.7.4

# LLM Integration
openai>=1.6.1
ollama==0.1.7

# Speech Processing
openai-whisper==20231117
gtts==2.4.0
TTS==0.22.0

# Twilio Integration
twilio==8.10.0

# Utilities
requests==2.31.0
aiofiles==23.2.1
numpy==1.24.3
sentence-transformers==2.2.2 